requirements.txt
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
dnspython==1.16.0
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
numpy==1.19.5
packaging==20.9
pandas==1.2.1
pluggy==0.13.1
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.19.1
pytz==2020.4
PyYAML==5.4.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.11
zipp==3.4.1
